daemon subsytem cleanup
